Create `pylinter#_handle_force_color_no_color`
Use the `_is_env_set_and_non_empty`, and the logic decided in the issue,
to modify the reporter(s) list accordingly.

At most, one reporter should be modified - the one going to `stdout`.

Hooking before the `self.set_reporter` was deemed as the smallest incision.

For no particular reason, it was decided that
`_handle_force_color_no_color` would modify its input;
so there would be no need for re-assigning the input we want to modify anyway.

Signed-off-by: Stavros Ntentos <133706+stdedos@users.noreply.github.com>